Safari-609.1.11.1/LayoutTests/accessibility/css-table-ignored.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<style>
.layout-table { display:table; }
.layout-table .second-cell, .layout-table .third-cell, .layout-table .third-cell-no-border { display: table-cell; }
.layout-table .third-cell { border-left: 5px solid #f00; }
</style>
</head>

<body id="body">

<div id="content">
<div class="layout-table" id="table">
    <div class="first-cell a11y">Visually hidden div</div>
    <div class="second-cell">Left div</div>
    <div class="third-cell">Right div</div>
</div>
</div>

<p id="description"></p>
<div id="console"></div>

<script>
description("This tests that an element use display:table will not be exposed as an accessible table.");

if (window.accessibilityController) {
    var table = accessibilityController.accessibleElementById("table");
    shouldBeTrue("!table || !table.isValid");
    document.getElementById("content").style.visibility = "hidden";
}

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
