LayoutTests/accessibility/css-table-ignored-expected.txt

This tests that an element use display:table will not be exposed as an accessible table.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !table || !table.isValid is true
PASS successfullyParsed is true

TEST COMPLETE
